fix: keep imports even when no fields are used
In some cases, a name could be imported and only used by itself (e.g. passed
into a mocking function). If the other side had a default export, then this
would be fine, but if the other side had no default export, we would just leave
off any type of import name. Now, we instead switch the import to an `import *`
in that case.