fix: add all files after running fix-imports, not just decaffeinated files
jscodeshift discovers and converts and files importing a file that was run
through decaffeinate, so we don't have a nice list of files to `git add`. But we
already made sure at the start that the git worktree has no changes, so it
should be safe to just add all files.