feat: change the recommended config file format from JSON to JS (#37)
Comments are useful for config files, and tailing commas and various JS features
are nice-to-have in config files, so I'm changing the format to just use JS. The
main downside is you need a `module.exports` at the start, but that seems like
not a huge deal. I used eslint's approach here: resolve the file to an abolute
path and use the `require-uncached` module.

JSON is still supported in case anyone wants it, but it's no longer documented
or tested, and may be removed in a later major version bump.