fix: run the fix-imports script through babel as part of the build process
The fix-imports script worked when run from source (using `npm link`), but not
when bulk-decaffeinate was installed through a package. The issue is that
jscodeshift wasn't properly running babel (I think because the file was within
node_modules), so the simple fix is to just run babel myself and ship the
compiled version.