fix: show nicer-looking commit messages (#41)
Instead of just saying the number of files that were converted, we now say the
name of the first file and the number of other files. There's more smart stuff
that could be done like finding a common directory among the files or using the
args that were passed in originally, but I'll leave that until later.