feat: add a "land" command (#47)
This is especially useful for Phabricator workflows where you ideally want to
update every commit message to point to the code review. But it's also useful in
general for making it a little easier to push a change that includes independent
commits.

Unfortunately, tests for this command are hard because it runs `fetch` and
`arc`, so I'll punt on tests for now.