fix: pre-process JS files to determine which are eligible for fix-imports
This is a bit faster than doing a full JS parse on each file, and is also nice
in that it lets us know which files we need to add in git.

This also fixes a race condition where the nodegit "status" command wasn't
returning all files that should be committed.