feat: add conversion and checking for literate coffeescript files (#110)
We now discover files ending in .litcoffee or .coffee.md and we change that
extension to JS like with any other file. This required going through the
convert step and fixing all of the places where it hard-coded .coffee.

In `check` and when verifying at the start of `convert`, we need to pass in the
`--literate` flag explicitly since we're just using stdin.