feat: rework handling of untracked files
Fixes #98
Fixes #120

Untracked files no longer block bulk-decaffeinate from running. But they do show
a warning since it might indicate a mistake.

Similarly, when discovering files, we skip any untracked files that would
otherwise be converted, and we show a warning in that case (hopefully untracked
CoffeeScript files are rare anyway).

If an untracked files are explicitly specified for conversion, we now fail with
a friendlier error than before.